requirements.txt
langgraph
langchain
langchain-groq
langchain-community
duckduckgo-search
beautifulsoup4
python-dotenv
streamlit
